Fix compiler warnings on Win64.
TEST=cctest/test-types
TBR=rossberg@chromium.org

Review URL: https://codereview.chromium.org/232773002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@20643 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/test-types.cc
@@ -73,8 +73,8 @@ class Types {
     }
 
     while (types.size() < kMaxTypes) {
-      size_t i = rng.NextInt(types.size());
-      size_t j = rng.NextInt(types.size());
+      size_t i = rng.NextInt(static_cast<int>(types.size()));
+      size_t j = rng.NextInt(static_cast<int>(types.size()));
       if (i != j) types.push_back(Type::Union(types[i], types[j], region));
     }
   }
